FFT online

+++ content/numerical/NumberTheoreticTransform.h
@@ -16,9 +16,8 @@
  */
 #pragma once
 
-#include "../number-theory/ModPow.h"
-
 const ll mod = (119 << 23) + 1, root = 62; // = 998244353
+#include "../number-theory/ModPow.h"
 // For p < 2^30 there is also e.g. 5 << 25, 7 << 26, 479 << 21
 // and 483 << 21 (same root). The last two are > 10^9.
 // int128: (2147483641LL<<32) - but 2xll & crt is faster.
@@ -48,7 +47,7 @@ vl conv(const vl &a, const vl &b) {
 	vl L(a), R(b), out(n);
 	L.resize(n), R.resize(n);
 	ntt(L), ntt(R);
-	rep(i,n) out[-i & (n - 1)] = (ll)L[i] * R[i] % mod * inv % mod;
+	rep(i,n) out[-i & (n - 1)] = L[i] * R[i] % mod * inv % mod;
 	ntt(out);
 	return {out.begin(), out.begin() + s};
 }
